Convert js to ts format
