+++ package.json
@@ -4,8 +4,8 @@
   "description": "Back-End Repo for Team 1 of Baboon/Bald Eagle Practicum",
   "main": "app.js",
   "scripts": {
-    "start": "node src/server.js",
-    "dev": "nodemon src/server.js",
+    "start": "ts-node src/server.ts",
+    "dev": "nodemon src/server.ts",
     "lint": "eslint src --report-unused-disable-directives --max-warnings 0",
     "lint:fix": "npm run lint -- --fix",
     "prettier": "npx prettier src --check",
