--- src/controllers/mainController.js
@@ -1,9 +1,0 @@
-const mainController = {}
-
-mainController.get = (req, res) => {
-  return res.json({
-    data: 'This is a full stack app!',
-  })
-}
-
-module.exports = mainController
