+++ src/routes/mainRouter.ts
@@ -1,0 +1,8 @@
+import express from 'express'
+import mainController from '../controllers/mainController'
+
+const router = express.Router()
+
+router.get('/', mainController)
+
+export default router
