+++ src/server.ts
@@ -1,5 +1,5 @@
 const { PORT = 8000 } = process.env
-const app = require('./app')
+import app from './app'
 
 const listener = () => console.log(`Listening on Port ${PORT}!`)
 app.listen(PORT, listener)
