--- src/test.js
@@ -1,1 +1,0 @@
-// chatbot github 2 slack test
